fix codegen

+++ test/cpp/jit/test_gpu.cpp
@@ -10048,7 +10048,7 @@ TEST(NVFuserTest, FusionMultiScanOp_CUDA) {
   std::vector<TensorView*> tvs = MultiScan(
       {BinaryOpType::Add, BinaryOpType::Max},
       {1, 2},
-      {new Float(0), new Float(0)},
+      {new Double(0), new Double(0)},
       tv0);
   auto tv2 = tvs[0];
   auto tv3 = tvs[1];
